+++ magda-web-client/src/Components/Dataset/Add/DatasetAddCommon.ts
@@ -1140,12 +1140,14 @@ async function convertStateToDatasetRecord(
         }));
     }
 
+    const authPolicy = authnReadPolicyId
+        ? authnReadPolicyId
+        : DEFAULT_POLICY_ID;
+
     const inputDataset = {
         id: datasetId,
         name: dataset.title,
-        authnReadPolicyId: authnReadPolicyId
-            ? authnReadPolicyId
-            : DEFAULT_POLICY_ID,
+        authnReadPolicyId: authPolicy,
         aspects: {
             publishing: getPublishingAspectData(state),
             "dcat-dataset-strings": buildDcatDatasetStrings(dataset),
@@ -1201,9 +1203,15 @@ async function convertStateToDatasetRecord(
     return inputDataset;
 }
 
-async function convertStateToDistributionRecords(state: State) {
+async function convertStateToDistributionRecords(
+    state: State,
+    authnReadPolicyId?: string
+) {
     const { dataset, distributions, licenseLevel } = state;
 
+    const authPolicy = authnReadPolicyId
+        ? authnReadPolicyId
+        : DEFAULT_POLICY_ID;
     const distributionRecords = distributions.map((distribution) => {
         const aspect =
             licenseLevel === "dataset"
@@ -1218,7 +1226,8 @@ async function convertStateToDistributionRecords(state: State) {
             name: distribution.title,
             aspects: {
                 "dcat-distribution-strings": aspect
-            }
+            },
+            authnReadPolicyId: authPolicy
         };
     });
 
@@ -1240,7 +1249,7 @@ export async function createDatasetFromState(
     }
 
     const distributionRecords = await (
-        await convertStateToDistributionRecords(state)
+        await convertStateToDistributionRecords(state, authnReadPolicyId)
     ).map((item) => {
         // --- set distribution initial version
         // --- the version will be bumped when it's superseded by a new file / distribution
@@ -1277,7 +1286,10 @@ export async function updateDatasetFromState(
 
     state.ckanExport[config.defaultCkanServer].exportRequired = true;
 
-    const distributionRecords = await convertStateToDistributionRecords(state);
+    const distributionRecords = await convertStateToDistributionRecords(
+        state,
+        authnReadPolicyId
+    );
     const datasetRecord = await convertStateToDatasetRecord(
         datasetId,
         distributionRecords,